fix typos and add middlewares
