+++ src/controllers/Auth.js
@@ -1,11 +1,10 @@
 /* eslint-disable prefer-destructuring */
 require('dotenv').config();
 const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
 const User = require('../models/user');
 
 // ******************************************************************* SignIn controller
-const SignIn = async (req, res) => {
+const signIn = async (req, res) => {
   try {
     const user = await User.findOne({ email: req.body.email });
 
@@ -28,7 +27,7 @@ const SignIn = async (req, res) => {
 };
 
 // ******************************************************************* SignUp controller
-const SignUp = async (req, res) => {
+const signUp = async (req, res) => {
   const { username, firstName, lastName, email, password, password2 } =
     req.body;
   try {
@@ -106,7 +105,7 @@ const cookSignUp = async (req, res) => {
 
 // ********************************************************** signout controller
 
-const SignOut = (req, res) => {
+const signOut = (req, res) => {
   try {
     // implement Sign out
     res.status(200).send({ message: 'Logged out successfully' });
@@ -115,59 +114,9 @@ const SignOut = (req, res) => {
   }
 };
 
-// ******************************************************* Authentication protectRoute controller
-const protectRoute = async (req, res, next) => {
-  // 1.getting token and check if it's there
-  let token;
-  if (
-    req.headers.authorization &&
-    req.headers.authorization.startsWith('Bearer')
-  ) {
-    token = req.headers.authorization.split(' ')[1];
-  }
-  if (!token) {
-    return res
-      .status(403)
-      .send({ message: 'you are not logged in, please log in to get access' });
-  }
-  // 2. verification token
-  const decoded = jwt.verify(token, process.env.JWT_SECRET);
-  // 3. check if user still exist
-  const freshUser = await User.findById(decoded.id);
-  if (!freshUser) {
-    return res
-      .status(401)
-      .send({ message: "the user of this token doesn't exists" });
-  }
-  // 4. check if the user changed password after the token is issued
-  if (freshUser.changedPasswordAfter(decoded.iat)) {
-    return res
-      .status(401)
-      .send({ message: 'user recently changed password, log in again' });
-  }
-
-  // grant access to protected route
-  req.user = freshUser;
-  return next();
-};
-
-// ******************************************************* Authoraization ristrictTo controller
-const ristrictTo =
-  (...type) =>
-  (req, res, next) => {
-    if (!type.includes(req.user.type)) {
-      return res
-        .status(403) // forbidden
-        .send({ message: 'you do not have permission to perform this action' });
-    }
-    return next();
-  };
-
 module.exports = {
-  SignIn,
-  SignUp,
-  protectRoute,
-  ristrictTo,
+  signIn,
+  signUp,
   cookSignUp,
-  SignOut,
+  signOut,
 };
