+++ src/routes/user.js
@@ -1,18 +1,20 @@
 const express = require('express');
 const AuthController = require('../controllers/Auth');
+const AuthMiddleware = require('../middlewares/auth');
 const userController = require('../controllers/user');
 
 const router = express.Router();
 // handling Authentication Routes
-router.post('/signin', AuthController.SignIn);
-router.post('/signup', AuthController.SignUp);
+router.post('/signin', AuthController.signIn);
+router.post('/signup', AuthController.signUp);
 router.post('/cook/signup', AuthController.cookSignUp);
-router.get('/signout', AuthController.SignOut);
+router.get('/signout', AuthController.signOut);
 // User Controllers Routes
+
 router.get(
   '/users',
-  AuthController.protectRoute,
-  AuthController.ristrictTo('admin'),
+  AuthMiddleware.protectRoute,
+  AuthMiddleware.ristrictTo('admin'),
   userController.getAllUsers
 );
 // router.get('/profile/:id', userController.getUserProfile);
